+++ src/main/java/driver/WebDriverManager.java
@@ -1,4 +1,4 @@
-package core;
+package driver;
 
 import org.openqa.selenium.WebDriver;
