+++ src/main/java/utils/EditTextUtils.java
@@ -1,0 +1,5 @@
+package utils;
+
+public class EditTextUtils {
+
+}
