+++ src/main/java/utils/Property.java
@@ -13,7 +13,7 @@ public class Property {
     public static String getPropertyValue(String propertyName) {
         Properties properties = new Properties();
 
-        try (InputStream inputStream = new FileInputStream("src/test/java/utils/config.properties")) {
+        try (InputStream inputStream = new FileInputStream("src/main/java/utils/config.properties")) {
             properties.load(inputStream);
         } catch (IOException e) {
             throw new UncheckedIOException(e);
